python/requirements.txt
numpy
Pillow
scikit-image
tqdm
torch
# For optional GPU support, install CuPy matching your CUDA version.
# Example for CUDA 11.x:
# cupy-cuda11x
